{"schema":"decodex/commit/1","summary":"Modularize host FFI frozen overlay bridge","authority":"manual"}
